+++ app/src/layers/components/LayerPanel.jsx
@@ -4,10 +4,10 @@ import LayerItem from 'layers/containers/LayerItem';
 import { LAYER_TYPES } from 'constants';
 import classnames from 'classnames';
 import ExpandItem from 'components/Shared/ExpandItem';
-import CarouselHeader from 'components/Shared/CarouselHeader';
+import AccordionHeader from 'components/Shared/AccordionHeader';
 import BasemapPanel from 'basemap/containers/BasemapPanel';
 import LayerListStyles from 'styles/components/map/item-list.scss';
-import CarouselStyles from 'styles/components/shared/carousel.scss';
+import AccordionStyles from 'styles/components/shared/accordion.scss';
 
 class LayerPanel extends Component {
   constructor(props) {
@@ -64,31 +64,31 @@ class LayerPanel extends Component {
     });
 
     return (
-      <div className={CarouselStyles.carousel}>
-        <CarouselHeader
+      <div className={AccordionStyles.accordion}>
+        <AccordionHeader
           menuName={'Basemaps'}
           openMenu={this.openMenu}
           expandState={this.state.expand}
         />
-        <ExpandItem active={this.state.expand === 'BASEMAPS'} carousel >
+        <ExpandItem active={this.state.expand === 'BASEMAPS'} accordion >
           <BasemapPanel />
         </ExpandItem >
-        <CarouselHeader
+        <AccordionHeader
           menuName={'Fishing Layers'}
           openMenu={this.openMenu}
           expandState={this.state.expand}
         />
-        <ExpandItem active={this.state.expand === 'FISHING_LAYERS'} carousel >
+        <ExpandItem active={this.state.expand === 'FISHING_LAYERS'} accordion >
           <ul className={LayerListStyles.list} >
             {fishingLayers}
           </ul >
         </ExpandItem >
-        <CarouselHeader
+        <AccordionHeader
           menuName={'Map Layers'}
           openMenu={this.openMenu}
           expandState={this.state.expand}
         />
-        <ExpandItem active={this.state.expand === 'MAP_LAYERS'} carousel >
+        <ExpandItem active={this.state.expand === 'MAP_LAYERS'} accordion >
           <ul className={classnames(LayerListStyles.list, LayerListStyles.shadow)} >
             {mapLayers}
           </ul >
